Update smartmeter paths
