+++ pyproject.toml
@@ -54,7 +54,7 @@ source = ["src", "*/site-packages"]
 
 [tool.coverage.run]
 branch = true
-source = ["smartmeter"]
+source = ["vienna_smartmeter"]
 
 [tool.coverage.report]
 show_missing = true
